Time: 103 ms (100.00%) | Memory: 48 MB (100.00%) - LeetSync

+++ 2711-minimum-time-to-visit-a-cell-in-a-grid/minimum-time-to-visit-a-cell-in-a-grid.cpp
@@ -23,7 +23,7 @@ public:
             if (dist[cell]<currentDist) continue;
             i = 0;
             while(i < 4) {
-                short nx = x + d[i], ny = y + d[++i];
+                int nx = x + d[i], ny = y + d[++i];
                 if (nx < 0 || ny < 0 || nx >= m || ny >= n)
                     continue;
                 int c_pos = nx + ny * m;
